requirements.txt
Pillow==10.1.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers==4.44.2
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
decord
moviepy
flash_attn>=1.0.0
flask>=2.0.1
flask_socketio>=5.3.0
eventlet>=0.33.0
nest_asyncio>=1.5.1
pyngrok>=5.0.0